requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.12.15
    # via langchain-community
aiosignal==1.4.0
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.10.0
    # via
    #   groq
    #   httpx
    #   watchfiles
attrs==25.3.0
    # via
    #   aiohttp
    #   jsonschema
    #   referencing
backoff==2.2.1
    # via posthog
bcrypt==4.3.0
    # via chromadb
build==1.3.0
    # via chromadb
cachetools==5.5.2
    # via google-auth
certifi==2025.8.3
    # via
    #   httpcore
    #   httpx
    #   kubernetes
    #   requests
charset-normalizer==3.4.2
    # via requests
chromadb==1.0.15
    # via langchain-chroma
click==8.2.1
    # via
    #   typer
    #   uvicorn
colorama==0.4.6
    # via
    #   build
    #   click
    #   tqdm
    #   uvicorn
coloredlogs==15.0.1
    # via onnxruntime
dataclasses-json==0.6.7
    # via langchain-community
distro==1.9.0
    # via
    #   groq
    #   posthog
durationpy==0.10
    # via kubernetes
filelock==3.18.0
    # via huggingface-hub
flatbuffers==25.2.10
    # via onnxruntime
frozenlist==1.7.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2025.7.0
    # via huggingface-hub
google-auth==2.40.3
    # via kubernetes
googleapis-common-protos==1.70.0
    # via opentelemetry-exporter-otlp-proto-grpc
greenlet==3.2.3
    # via sqlalchemy
groq==0.30.0
    # via langchain-groq
grpcio==1.74.0
    # via
    #   chromadb
    #   opentelemetry-exporter-otlp-proto-grpc
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via
    #   chromadb
    #   groq
    #   langsmith
httpx-sse==0.4.1
    # via langchain-community
huggingface-hub==0.34.3
    # via
    #   langchain-huggingface
    #   tokenizers
humanfriendly==10.0
    # via coloredlogs
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
importlib-metadata==8.7.0
    # via opentelemetry-api
importlib-resources==6.5.2
    # via chromadb
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
jsonschema==4.25.0
    # via chromadb
jsonschema-specifications==2025.4.1
    # via jsonschema
kubernetes==33.1.0
    # via chromadb
langchain==0.3.27
    # via
    #   pdf-rag-learning (pyproject.toml)
    #   langchain-community
langchain-chroma==0.2.5
    # via pdf-rag-learning (pyproject.toml)
langchain-community==0.3.27
    # via pdf-rag-learning (pyproject.toml)
langchain-core==0.3.72
    # via
    #   langchain
    #   langchain-chroma
    #   langchain-community
    #   langchain-groq
    #   langchain-huggingface
    #   langchain-text-splitters
langchain-groq==0.3.6
    # via pdf-rag-learning (pyproject.toml)
langchain-huggingface==0.3.1
    # via pdf-rag-learning (pyproject.toml)
langchain-text-splitters==0.3.9
    # via langchain
langsmith==0.4.11
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
markdown-it-py==3.0.0
    # via rich
marshmallow==3.26.1
    # via dataclasses-json
mdurl==0.1.2
    # via markdown-it-py
mmh3==5.2.0
    # via chromadb
mpmath==1.3.0
    # via sympy
multidict==6.6.3
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.1.0
    # via typing-inspect
numpy==2.3.2
    # via
    #   chromadb
    #   langchain-chroma
    #   langchain-community
    #   onnxruntime
oauthlib==3.3.1
    # via
    #   kubernetes
    #   requests-oauthlib
onnxruntime==1.22.1
    # via chromadb
opentelemetry-api==1.36.0
    # via
    #   chromadb
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
opentelemetry-exporter-otlp-proto-common==1.36.0
    # via opentelemetry-exporter-otlp-proto-grpc
opentelemetry-exporter-otlp-proto-grpc==1.36.0
    # via chromadb
opentelemetry-proto==1.36.0
    # via
    #   opentelemetry-exporter-otlp-proto-common
    #   opentelemetry-exporter-otlp-proto-grpc
opentelemetry-sdk==1.36.0
    # via
    #   chromadb
    #   opentelemetry-exporter-otlp-proto-grpc
opentelemetry-semantic-conventions==0.57b0
    # via opentelemetry-sdk
orjson==3.11.1
    # via
    #   chromadb
    #   langsmith
overrides==7.7.0
    # via chromadb
packaging==25.0
    # via
    #   build
    #   huggingface-hub
    #   langchain-core
    #   langsmith
    #   marshmallow
    #   onnxruntime
posthog==5.4.0
    # via chromadb
propcache==0.3.2
    # via
    #   aiohttp
    #   yarl
protobuf==6.31.1
    # via
    #   googleapis-common-protos
    #   onnxruntime
    #   opentelemetry-proto
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pybase64==1.4.2
    # via chromadb
pydantic==2.11.7
    # via
    #   chromadb
    #   groq
    #   langchain
    #   langchain-core
    #   langsmith
    #   pydantic-settings
pydantic-core==2.33.2
    # via pydantic
pydantic-settings==2.10.1
    # via langchain-community
pygments==2.19.2
    # via rich
pypdf==5.9.0
    # via pdf-rag-learning (pyproject.toml)
pypika==0.48.9
    # via chromadb
pyproject-hooks==1.2.0
    # via build
pyreadline3==3.5.4
    # via humanfriendly
python-dateutil==2.9.0.post0
    # via
    #   kubernetes
    #   posthog
python-dotenv==1.1.1
    # via
    #   pdf-rag-learning (pyproject.toml)
    #   pydantic-settings
    #   uvicorn
pyyaml==6.0.2
    # via
    #   chromadb
    #   huggingface-hub
    #   kubernetes
    #   langchain
    #   langchain-community
    #   langchain-core
    #   uvicorn
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.4
    # via
    #   huggingface-hub
    #   kubernetes
    #   langchain
    #   langchain-community
    #   langsmith
    #   posthog
    #   requests-oauthlib
    #   requests-toolbelt
requests-oauthlib==2.0.0
    # via kubernetes
requests-toolbelt==1.0.0
    # via langsmith
rich==14.1.0
    # via
    #   chromadb
    #   typer
rpds-py==0.26.0
    # via
    #   jsonschema
    #   referencing
rsa==4.9.1
    # via google-auth
shellingham==1.5.4
    # via typer
six==1.17.0
    # via
    #   kubernetes
    #   posthog
    #   python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   groq
sqlalchemy==2.0.42
    # via
    #   langchain
    #   langchain-community
sympy==1.14.0
    # via onnxruntime
tenacity==9.1.2
    # via
    #   chromadb
    #   langchain-community
    #   langchain-core
tokenizers==0.21.4
    # via
    #   chromadb
    #   langchain-huggingface
tqdm==4.67.1
    # via
    #   chromadb
    #   huggingface-hub
typer==0.16.0
    # via chromadb
typing-extensions==4.14.1
    # via
    #   chromadb
    #   groq
    #   huggingface-hub
    #   langchain-core
    #   opentelemetry-api
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typer
    #   typing-inspect
    #   typing-inspection
typing-inspect==0.9.0
    # via dataclasses-json
typing-inspection==0.4.1
    # via
    #   pydantic
    #   pydantic-settings
urllib3==2.5.0
    # via
    #   kubernetes
    #   requests
uvicorn==0.35.0
    # via chromadb
watchfiles==1.1.0
    # via uvicorn
websocket-client==1.8.0
    # via kubernetes
websockets==15.0.1
    # via uvicorn
yarl==1.20.1
    # via aiohttp
zipp==3.23.0
    # via importlib-metadata
zstandard==0.23.0
    # via langsmith
